client/pyproject.toml
[project]
name = "skilltool"
version = "0.1.0"
description = "PyPI-like registry client for Claude Agent skill.md packages"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Ryo Nakagami" }]
keywords = ["skill", "claude", "agent", "registry", "pypi"]
dependencies = [
    "typer>=0.12",
    "httpx>=0.27",
    "rich>=13.7",
    "pyyaml>=6.0",
]

[project.scripts]
skilltool = "skilltool.cli:app"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/skilltool"]
